+++ src/main/java/io/xpring/javascript/JavaScriptSigner.java
@@ -4,8 +4,8 @@ import io.xpring.proto.Transaction;
 import io.xpring.proto.SignedTransaction;
 
 import com.google.protobuf.InvalidProtocolBufferException;
-import io.xpring.Utils;
-import io.xpring.Wallet;
+import io.xpring.xrpl.Utils;
+import io.xpring.xrpl.Wallet;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Value;
