+++ src/test/java/io/xpring/xrpl/WalletTest.java
@@ -1,4 +1,4 @@
-package io.xpring;
+package io.xpring.xrpl;
 
 import org.junit.Rule;
 import org.junit.Test;
